test: move bloom_filter unit tests to tests/
Relocate the 4 #[cfg(test)] tests from src/bloom_filter.rs to
tests/bloom_filter.rs, matching the tests-live-in-tests/ convention. They
exercise only the public BloomFilter API (new/insert/might_contain), so the
move is mechanical (use ironbeam::bloom_filter::BloomFilter).

src/planner.rs keeps its inline `dominator_tests`: they test
`build_dominator_tree` (private) and `find_cache_node_via_dominators`
(pub(crate)), which an integration test cannot reach. Per the plan's G5,
unit tests of private internals legitimately stay in-module; the planner's
public behavior is already covered by tests/planner.rs.
